remove AsNoTracking() calls
for the sake of simplicity, and stop the madness that arisen from unit testing calls that use this method (couldn't find explanation for the resulting behavior of some 'Contains' calls), these calls will be removed from the classes. Performance is not yet an issue for this to be used in the project.